crm-ai-task/backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
langgraph
langchain
langchain-groq
groq
python-dotenv
